pyproject.toml
[tool.poetry]
name = "graffitihouse"
version = "0.1.0"
description = "Web application for Graffiti House"
authors = ["Jason Heppler <jason.heppler@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.1.7"
django-extensions = "^3.2.1"
Pillow = "^9.4.0"
django-tailwind = "^3.3.0"
django-browser-reload = "^1.7.0"
psycopg2 = "^2.9.5"
python-dotenv = "^1.0.0"
django-environ = "^0.10.0"
django-prose = "^1.2.0"
django-admin-interface = "^0.24.2"
django-taggit = "^4.0.0"
taggit-selectize = "^2.11.0"
django-multiselectfield = "^0.1.12"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
djhtml = "^3.0.6"
pre-commit = "^3.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
